requirements.txt
streamlit
pandas
python-dotenv
psycopg2-binary
boto3
openpyxl
